Change the owner and permissions to the html synced folder
Change the owner of /var/www/html back to the vagrant user and adjust the folder permissions so www-data still has write access via the group. This fixes the occasional nuisance of working inside the vagrant instance and having to sudo for commands that shouldn't really be run using it.